+++ Src/Usr_adc.h
@@ -4,5 +4,5 @@
 void Adc_init(void);
 u16 Adc_Value_Get(void);
 
-extern u16	BatVoltage_Adc;
+extern u32	BatVoltage_Adc;
 #endif
